gatsby/cache-dir: Emit relative paths to core JS assets.
Early prototype for #14161

This is an early attempt at allowing sites to be hosted at any
arbitrary location, with a possibly unknown site prefix. Further
changes are required to optionally avoid page redirects when opened
from a local filesystem, as well as to ensure that gatsby-link can
also emit relative paths and append an index.html.